Consolidate Oracle Platforms
* drop specific support for Oracle DB 9 and older
