+++ foundation/org.eclipse.persistence.core/src/main/java/org/eclipse/persistence/platform/database/Oracle11Platform.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2019 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2021 Oracle and/or its affiliates. All rights reserved.
  * Copyright (c) 2019 IBM Corporation. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -24,7 +24,7 @@ package org.eclipse.persistence.platform.database;
  * Provides Oracle version specific behavior when 
  * org.eclipse.persistence.oracle bundle is not available.
  */
-public class Oracle11Platform extends Oracle10Platform {
+public class Oracle11Platform extends OraclePlatform {
     public Oracle11Platform() {
         super();
