+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/framework/oracle/OracleDBPlatformHelper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -15,16 +15,18 @@
 package org.eclipse.persistence.testing.framework.oracle;
 
 import org.eclipse.persistence.platform.database.OraclePlatform;
-import org.eclipse.persistence.platform.database.oracle.Oracle8Platform;
-import org.eclipse.persistence.platform.database.oracle.Oracle9Platform;
+import org.eclipse.persistence.platform.database.oracle.Oracle11Platform;
+import org.eclipse.persistence.platform.database.oracle.Oracle12Platform;
 
 public class OracleDBPlatformHelper extends org.eclipse.persistence.testing.framework.OracleDBPlatformHelper{
 
+    @Override
     public OraclePlatform getOracle8Platform(){
-        return new Oracle8Platform();
+        return new Oracle11Platform();
     }
 
+    @Override
     public OraclePlatform getOracle9Platform(){
-        return new Oracle9Platform();
+        return new Oracle12Platform();
     }
 }
