+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/tests/conversion/DataTypesConvertedToAClassForOracle9Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,13 +14,13 @@
 //     Oracle - initial API and implementation from Oracle TopLink
 package org.eclipse.persistence.testing.tests.conversion;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.sql.Time;
+import java.sql.Timestamp;
 import java.util.Calendar;
 
-import java.sql.*;
-
-import java.math.*;
-
-import org.eclipse.persistence.platform.database.oracle.Oracle9Platform;
+import org.eclipse.persistence.platform.database.oracle.Oracle11Platform;
 import org.eclipse.persistence.testing.framework.TestWarningException;
 
 //This test retrieves all the classes that can be converted to a given class by
@@ -37,6 +37,7 @@ public class DataTypesConvertedToAClassForOracle9Test extends DataTypesConverted
         setDescription("Test getDataTypesConvertedTo() in Oracle9Platform.");
     }
 
+    @Override
     public void setup() {
         if(!getSession().getPlatform().isOracle()) {
             throw new TestWarningException("This test requires Oracle database");
@@ -44,8 +45,9 @@ public class DataTypesConvertedToAClassForOracle9Test extends DataTypesConverted
         cm = getSession().getPlatform();
     }
 
+    @Override
     protected boolean isChar(Class aClass) {
-        return super.isChar(aClass) || aClass == Oracle9Platform.NCHAR || aClass == Oracle9Platform.NSTRING ||
-            aClass == Oracle9Platform.NCLOB;
+        return super.isChar(aClass) || aClass == Oracle11Platform.NCHAR || aClass == Oracle11Platform.NSTRING ||
+            aClass == Oracle11Platform.NCLOB;
     }
 }
