+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/tests/nchar/BaseNcharTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,10 +14,13 @@
 //     Oracle - initial API and implementation from Oracle TopLink
 package org.eclipse.persistence.testing.tests.nchar;
 
-import org.eclipse.persistence.testing.framework.*;
-import org.eclipse.persistence.sessions.UnitOfWork;
 import org.eclipse.persistence.internal.databaseaccess.DatabasePlatform;
-import org.eclipse.persistence.platform.database.oracle.Oracle9Platform;
+import org.eclipse.persistence.platform.database.oracle.Oracle11Platform;
+import org.eclipse.persistence.sessions.UnitOfWork;
+import org.eclipse.persistence.testing.framework.TestCase;
+import org.eclipse.persistence.testing.framework.TestException;
+import org.eclipse.persistence.testing.framework.TestProblemException;
+import org.eclipse.persistence.testing.framework.TestWarningException;
 
 public abstract class BaseNcharTest extends TestCase {
     protected DatabasePlatform platformOriginal;
@@ -25,10 +28,11 @@ public abstract class BaseNcharTest extends TestCase {
     protected CharNchar controlObject;
     protected boolean usesStringBindingOriginal;
 
-    protected Oracle9Platform getOracle9Platform() throws ClassCastException {
-        return (Oracle9Platform)getSession().getPlatform();
+    protected Oracle11Platform getOracle9Platform() throws ClassCastException {
+        return (Oracle11Platform)getSession().getPlatform();
     }
 
+    @Override
     protected void setup() {
         if (!getSession().getPlatform().isOracle()) {
             throw new TestWarningException("This test case works on Oracle only");
@@ -38,7 +42,7 @@ public abstract class BaseNcharTest extends TestCase {
         } catch (ClassCastException ex) {
             DatabasePlatform platform = getSession().getPlatform();
             try {
-                getSession().getLogin().usePlatform(new org.eclipse.persistence.platform.database.oracle.Oracle9Platform());
+                getSession().getLogin().usePlatform(new org.eclipse.persistence.platform.database.oracle.Oracle11Platform());
                 getDatabaseSession().logout();
                 getDatabaseSession().login();
                 platformOriginal = platform;
@@ -55,6 +59,7 @@ public abstract class BaseNcharTest extends TestCase {
         }
     }
 
+    @Override
     protected void verify() {
         if (object == null) {
             throw new TestException("object is missing");
@@ -166,6 +171,7 @@ public abstract class BaseNcharTest extends TestCase {
         }
     }
 
+    @Override
     public void reset() {
         if (object != null) {
             UnitOfWork uow = getSession().acquireUnitOfWork();
