+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/tests/nonJDBC/NonJDBCTestHelper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -12,12 +12,12 @@
 
 package org.eclipse.persistence.testing.tests.nonJDBC;
 
+// JUnit imports
+import static org.junit.Assert.fail;
+
 // javase imports
 import java.lang.reflect.Field;
 import java.util.Properties;
-
-// JUnit imports
-import static org.junit.Assert.fail;
 
 // EclipseLink imports
 import org.eclipse.persistence.descriptors.RelationalDescriptor;
@@ -26,7 +26,7 @@ import org.eclipse.persistence.internal.sessions.factories.ObjectPersistenceWork
 import org.eclipse.persistence.mappings.DirectToFieldMapping;
 import org.eclipse.persistence.mappings.transformers.ConstantTransformer;
 import org.eclipse.persistence.oxm.mappings.XMLTransformationMapping;
-import org.eclipse.persistence.platform.database.oracle.Oracle10Platform;
+import org.eclipse.persistence.platform.database.oracle.Oracle11Platform;
 import org.eclipse.persistence.platform.xml.XMLComparer;
 import org.eclipse.persistence.platform.xml.XMLParser;
 import org.eclipse.persistence.platform.xml.XMLPlatformFactory;
@@ -81,7 +81,7 @@ public class NonJDBCTestHelper {
         login.setPassword(password);
         login.setConnectionString(url);
         login.setDriverClassName(driver);
-        login.setDatasourcePlatform(new Oracle10Platform());
+        login.setDatasourcePlatform(new Oracle11Platform());
         login.useBatchWriting();
         project.setDatasourceLogin(login);
         return project;
