+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/tests/plsqlrecord/PLSQLrecordWithCompatibleTypeInTestSet.java
@@ -10,7 +10,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
  */
 
-// Copyright (c) 2011, 2015 Oracle and/or its affiliates. All rights reserved.
+// Copyright (c) 2011, 2021 Oracle and/or its affiliates. All rights reserved.
 
 package org.eclipse.persistence.testing.tests.plsqlrecord;
 
@@ -36,8 +36,8 @@ import org.eclipse.persistence.internal.sessions.DatabaseSessionImpl;
 import org.eclipse.persistence.internal.sessions.factories.ObjectPersistenceWorkbenchXMLProject;
 import org.eclipse.persistence.oxm.XMLContext;
 import org.eclipse.persistence.oxm.XMLMarshaller;
+import org.eclipse.persistence.platform.database.OraclePlatform;
 import org.eclipse.persistence.platform.database.jdbc.JDBCTypes;
-import org.eclipse.persistence.platform.database.oracle.Oracle8Platform;
 import org.eclipse.persistence.platform.database.oracle.plsql.PLSQLStoredProcedureCall;
 import org.eclipse.persistence.platform.database.oracle.plsql.PLSQLrecord;
 import org.eclipse.persistence.queries.DataModifyQuery;
@@ -269,7 +269,7 @@ public class PLSQLrecordWithCompatibleTypeInTestSet  {
         };
         Struct struct = null;
         try {
-            struct = ((Oracle8Platform)s.getPlatform()).createStruct(
+            struct = ((OraclePlatform)s.getPlatform()).createStruct(
                 "EMP_TYPE", attributes, ((DatabaseSessionImpl)s).getAccessor().getConnection());
         }
         catch (SQLException e) {
