+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/tests/types/OracleTIMESTAMPTypeTestSystem.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,17 +14,18 @@
 //     Oracle - initial API and implementation from Oracle TopLink
 package org.eclipse.persistence.testing.tests.types;
 
-import org.eclipse.persistence.platform.database.oracle.Oracle9Platform;
+import org.eclipse.persistence.platform.database.oracle.Oracle11Platform;
 import org.eclipse.persistence.sessions.DatabaseSession;
 import org.eclipse.persistence.testing.framework.TestSystem;
 import org.eclipse.persistence.tools.schemaframework.SchemaManager;
 
 public class OracleTIMESTAMPTypeTestSystem extends TestSystem {
 
+    @Override
     public void createTables(DatabaseSession session) {
         SchemaManager schemaManager = new SchemaManager(session);
 
-        if (session.getPlatform() instanceof Oracle9Platform) {
+        if (session.getPlatform() instanceof Oracle11Platform) {
             schemaManager.replaceObject(TIMESTAMPDirectToFieldTester.tableDefinition(session));
             schemaManager.replaceObject(TIMESTAMPTypeConversionTester.tableDefinition(session));
             schemaManager.replaceObject(CalendarToTSTZWithoutSessionTZTest.tableDefinition(session));
