+++ foundation/org.eclipse.persistence.oracle.test/src/it/java/org/eclipse/persistence/testing/tests/xdb/ExistsNodeWithJoinTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,9 +14,11 @@
 //     Oracle - initial API and implementation from Oracle TopLink
 package org.eclipse.persistence.testing.tests.xdb;
 
-import org.eclipse.persistence.expressions.*;
-import org.eclipse.persistence.platform.database.oracle.Oracle9Platform;
-import org.eclipse.persistence.testing.framework.*;
+import org.eclipse.persistence.expressions.ExpressionBuilder;
+import org.eclipse.persistence.platform.database.oracle.Oracle11Platform;
+import org.eclipse.persistence.testing.framework.TestCase;
+import org.eclipse.persistence.testing.framework.TestErrorException;
+import org.eclipse.persistence.testing.framework.TestWarningException;
 
 public class ExistsNodeWithJoinTest extends TestCase {
     Employee_XML employee = null;
@@ -25,19 +27,23 @@ public class ExistsNodeWithJoinTest extends TestCase {
         setDescription("Tests the use of the existsNode() function as well as testing use of XMLType in a join");
     }
 
+    @Override
     public void setup() {
-        if (!(getSession().getPlatform() instanceof Oracle9Platform)) {
+        if (!(getSession().getPlatform() instanceof Oracle11Platform)) {
             throw new TestWarningException("This test is intended for the Oracle9Platform");
         }
     }
 
+    @Override
     public void reset() {
     }
 
+    @Override
     public void test() {
         employee = (Employee_XML)getSession().readObject(Employee_XML.class, new ExpressionBuilder().get("manager").get("resume").existsNode("//education/degree").greaterThan(0));
     }
 
+    @Override
     public void verify() {
         if ((employee == null) || !employee.firstName.equals("Frank")) {
             throw new TestErrorException("Wrong Employee returned:" + employee);
