+++ foundation/org.eclipse.persistence.oracle/src/main/java/org/eclipse/persistence/platform/database/oracle/Oracle12Platform.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -45,6 +45,7 @@ public class Oracle12Platform extends Oracle11Platform {
 
     public Oracle12Platform() {
         super();
+        usesNativeBatchWriting = false;
         supportsIdentity = true;
         identitySequences = new ConcurrentHashMap<>();
     }
